% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ave_value.R
\name{ave_value}
\alias{ave_value}
\title{The Average Value of Every State}
\usage{
ave_value(input_state)
}
\arguments{
\item{input_state}{character. Every state of US from HoneyProduction dataset.}
}
\value{
A tibble of dataset that contained the state name and mean honey value of that state.
}
\description{
This is a function that can calculate the average honey value
that created by every state in US.
}
\examples{
ave_value("AL")
ave_value(input_state = "AL")

}
