% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aveprice.R
\name{price_plot}
\alias{price_plot}
\title{Plotting Average Price of Every State}
\usage{
price_plot(state_name)
}
\arguments{
\item{state_name}{character. Every state of US from HoneyProduction dataset.}
}
\value{
A bar chart showing the average price of honey in each state,
with selected states highlighted and others grayed out,
meanwhile each state's price data displayed on its own bar.
}
\description{
This function takes a `state' object to plot the average price histogram for the input state,
and this value can be easily found from the graph by highlighting the selected state.
}
\examples{
price_plot("AL")
price_plot(state_name = "AL")

}
