% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_production.R
\name{production_plot}
\alias{production_plot}
\title{Plotting Total Production Trend From 1998 to 2012}
\usage{
production_plot(allstates)
}
\arguments{
\item{allstates}{character. Every state of US from HoneyProduction dataset.}
}
\value{
A line graph containing the curve of the change
in the total honey production of each state from 1998 to 2012.
}
\description{
A line graph showing total honey production trend by states in the US over time,
which is a big picture of the honey production data.

The output is the line graph of the honey production value for all states,
regardless of whether argument has an input or not.
}
\examples{
production_plot()
production_plot(allstates = )

}
