% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shinyapp.R
\name{run_app}
\alias{run_app}
\title{Honey Production Shiny Application}
\usage{
run_app()
}
\value{
A shiny app. By selecting different years, interactively showing the changing trend of honey production in every state of USA during 1998 to 2012. Showing the mean value of honey price per pound of states, and visualizing the top 6 states with highest value of honey of the data.
}
\description{
The app uses an interactive component that allows users to self-select a year
or state to see the impact of different choices on yield or average unit price.
It also presents the top six states with the highest yields to the user through a data visualization.
}
