% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_production.R
\name{yearly_production}
\alias{yearly_production}
\title{Summary of Total Honey Production of US States(1998-2012)}
\usage{
yearly_production(input_year)
}
\arguments{
\item{input_year}{integer. Includes all years from 1998 to 2012.}
}
\value{
A dataframe containing summary values (i.e. minimum, Q1, median, Q3, maximum)
of the total production for the selected year.
}
\description{
The function is used for showing the summary information of
total production of US in the period of 1998 to 2012.
}
\examples{
yearly_production(2000)
yearly_production(input_year = 2000)

}
